backend/requirements.txt
Pillow
pyzbar
requests